# Conflicts:
#	.github/workflows/native-unit-test.yml
#	.vscode/settings.json
#	native/CMakeLists.txt
#	native/cmake/predefine.cmake
#	native/cocos/audio/android/AudioDecoderSLES.h
#	native/cocos/audio/android/AudioEngine-inl.cpp
#	native/cocos/audio/android/AudioPlayerProvider.cpp
#	native/cocos/audio/android/AudioResampler.h
#	native/cocos/bindings/auto/jsb_cocos_auto.cpp
#	native/cocos/bindings/auto/jsb_cocos_auto.h
#	native/cocos/bindings/jswrapper/config.h
#	native/cocos/bindings/manual/jsb_conversions.h
#	native/cocos/bindings/manual/jsb_dragonbones_manual.cpp
#	native/cocos/bindings/manual/jsb_global.cpp
#	native/cocos/bindings/manual/jsb_module_register.cpp
#	native/cocos/engine/Engine.cpp
#	native/cocos/platform/interfaces/modules/Device.cpp
#	native/extensions/assets-manager/AssetsManagerEx.cpp
#	native/external-config.json
#	platforms/native/engine/jsb-loader.js
#	platforms/native/engine/jsb-scene.js
#	scripts/build-engine/package-lock.json
#	scripts/build-engine/package.json
#	templates/cmake/openharmony.cmake

+++ native/cocos/audio/android/AudioResampler.h
@@ -17,10 +17,10 @@
 #pragma once
 
 #include <stdint.h>
-#include <sys/types.h>
 #if CC_PLATFORM == CC_PLATFORM_ANDROID
 #include <android/log.h>
 #include <sys/system_properties.h>
+#include <sys/types.h>
 #endif
 
 #include "audio/android/AudioBufferProvider.h"
@@ -151,7 +151,7 @@ protected:
     //  phaseWrapLimit is the wraparound (1 << kNumPhaseBits), if not specified explicitly.
     //
     inline size_t getInFrameCountRequired(size_t outFrameCount) {
-        return (static_cast<uint64_t>(outFrameCount) * mInSampleRate + (mSampleRate - 1)) / mSampleRate;
+        return (static_cast<size_t>(outFrameCount) * mInSampleRate + (mSampleRate - 1)) / mSampleRate;
     }
 
     inline float clampFloatVol(float volume) {
